Add old temporary model.py file.
These files uses a much older version of pabutools to run which looking at the current version, is depreciated. Rather than just modify it so it works with the newest version, since that takes time we may not have and we're only focusing on the HTML. We add the model.py file from the old pabutools version to this directory which contains all the classes required to compile this successfully.