Extract initial and final voter funding variables.
Using the same record_round_end method, we add a new parameter 'voters' to extract information about each voter's budget during a round. This is used to find the total funds the voters had to vote for the winning project (i.e. final_voter_funding).
Whereas, we get the initial budget per voter and multiply with the number of voters who voted for the winning project to get the initial_voter_funding.