Connect render function to page summary template.
Connect the render function from MESDataStore to the page summary template to generate a page summary for any election. This uses the values extracted from the previous commit to create the charts shown in the summary table. However, there are some issues worth noting, the selected rounds doesn't seem to be matching the selected projects in the outcome, further, the effective vote seems to be incorrect.