Connect summary table with round by round analysis.
Connected summary table with the round by round analysis page, so when clicking a row, the row expands and the user can click on the hyperlink and jump to the specific round in the analysis page. Further, removed summary statistics carousel from round by round analysis page and move to summary table page.