Finish summay explanations and minor fix UI issues.
Completed the summary explanations within each row, these may need changing or adding but that can be done later, for now this is fine. Further, fixed the blue bar tippy tool tip for when it is highlighted for a rejected projects. Finally, added some general explanations on top of the page for the budget chart and the summary table.